+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/versionspecific/SequencedCollectionsHelper.java
@@ -1,0 +1,10 @@
+package nl.jqno.equalsverifier.internal.versionspecific;
+
+import nl.jqno.equalsverifier.internal.instantiation.vintage.FactoryCache;
+
+public final class SequencedCollectionsHelper {
+
+    private SequencedCollectionsHelper() {}
+
+    public static void add(FactoryCache factoryCache) {}
+}