+++ equalsverifier-release-verify/src/test/java/nl/jqno/equalsverifier/verify_release/jar/helper/JarAsserter.java
@@ -32,7 +32,8 @@ public class JarAsserter {
         assertPresenceOf(
             "/META-INF/versions/11" + EV + "/internal/versionspecific/ModuleHelper.class",
             "/META-INF/versions/16" + EV + "/internal/versionspecific/RecordsHelper.class",
-            "/META-INF/versions/17" + EV + "/internal/versionspecific/SealedTypesHelper.class");
+            "/META-INF/versions/17" + EV + "/internal/versionspecific/SealedTypesHelper.class",
+            "/META-INF/versions/21" + EV + "/internal/versionspecific/SequencedCollectionsHelper.class");
     }
 
     public void assertPresenceOfEmbeddedDepedencies() {
@@ -94,7 +95,10 @@ public class JarAsserter {
                 "/META-INF/versions/16" + EV + "/internal/versionspecific/RecordsHelper.class"),
             () -> assertVersionOfClassFile(
                 61,
-                "/META-INF/versions/17" + EV + "/internal/versionspecific/SealedTypesHelper.class"));
+                "/META-INF/versions/17" + EV + "/internal/versionspecific/SealedTypesHelper.class"),
+            () -> assertVersionOfClassFile(
+                65,
+                "/META-INF/versions/21" + EV + "/internal/versionspecific/SequencedCollectionsHelper.class"));
     }
 
     public void assertVersionsOfEmbeddedClassFiles() {